requirements-dev.txt

black
flake8
numexpr
numpy
pandas
pre-commit
pytest
pyyaml
scikit-learn
uproot>=4.0
